+++ src/main/java/ynch/ui/Deadline.java
@@ -1,3 +1,5 @@
+package ynch.ui;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
