+++ src/main/java/ynch/ui/DialogBox.java
@@ -1,0 +1,20 @@
+package ynch.ui;
+
+import javafx.geometry.Pos;
+import javafx.scene.control.Label;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+
+public class DialogBox extends HBox {
+
+    private Label text;
+    private ImageView displayPicture;
+
+    public DialogBox(String s, Image i) {
+        text = new Label(s);
+        displayPicture = new ImageView(i);
+        this.getChildren().addAll(text, displayPicture);
+    }
+}
+
