+++ src/main/java/ynch/ui/EmptyTaskException.java
@@ -1,3 +1,5 @@
+package ynch.ui;
+
 public class EmptyTaskException extends Exception {
     public EmptyTaskException() {
         super("Meow!! Description of a todo cannot be empty :(");
