+++ src/main/java/ynch/ui/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package ynch.ui;
+
 public class InvalidCommandException extends Exception {
     public InvalidCommandException() {
         super("Meow!! I don't know what you're yapping about :(");
