+++ src/main/java/ynch/ui/Task.java
@@ -1,3 +1,5 @@
+package ynch.ui;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
