+++ src/main/java/ynch/ui/Todo.java
@@ -1,3 +1,5 @@
+package ynch.ui;
+
 /**
  * Represents a Todo task that extends the Task class.
  */
