+++ src/main/java/ynch/ui/ValidCommand.java
@@ -1,3 +1,5 @@
-enum ValidCommand {
+package ynch.ui;
+
+public enum ValidCommand {
     todo, deadline, event, mark, unmark, list, bye, delete, find
 }
