+++ src/main/java/ynch/ui/Ynch.txt
@@ -1,3 +1,4 @@
 D1/homework /2024-06-12
 E0/party all day /2024-11-09/2024-11-09
 D0/meowmeow /2025-05-01
+T0/meow
